Struct usb_util::Device

pub struct Device {
    fd: Arc<File>,
    device_descriptor_tree: DeviceDescriptorTree,
}

Device represents a USB device.

Fields
fd: Arc<File>device_descriptor_tree: DeviceDescriptorTree
Implementations

impl Device

pub fn new(fd: File) -> Result<Self>

Create a new Device from a file descriptor. fd should be a file in usbdevfs (e.g. /dev/bus/usb/001/002).

pub fn fd(&self) -> Arc<File>

unsafe fn ioctl(&self, nr: IoctlNr) -> Result<i32>

unsafe fn ioctl_with_ref<T>(&self, nr: IoctlNr, arg: &T) -> Result<i32>

unsafe fn ioctl_with_mut_ref<T>(&self, nr: IoctlNr, arg: &mut T) -> Result<i32>

unsafe fn ioctl_with_mut_ptr<T>(&self, nr: IoctlNr, arg: *mut T) -> Result<i32>

pub fn submit_transfer(&mut self, transfer: Transfer) -> Result<TransferHandle>

Submit a transfer to the device. The transfer will be processed asynchronously by the device. Call poll_transfers() on this device to check for completed transfers.

pub fn poll_transfers(&self) -> Result<()>

Check for completed asynchronous transfers submitted via submit_transfer(). The callback for each completed transfer will be called.

pub fn reset(&self) -> Result<()>

Perform a USB port reset to reinitialize a device.

pub fn claim_interface(&self, interface_number: u8) -> Result<()>

Claim an interface on this device.

pub fn release_interface(&self, interface_number: u8) -> Result<()>

Release an interface previously claimed with claim_interface().

pub fn set_interface_alt_setting(
&self,
interface_number: u8,
alternative_setting: u8
) -> Result<()>

Activate an alternate setting for an interface.

pub fn set_active_configuration(&mut self, config: u8) -> Result<()>

Set active configuration for this device.

pub fn get_device_descriptor(&self) -> Result<DeviceDescriptor>

Get the device descriptor of this device.

pub fn get_device_descriptor_tree(&self) -> &DeviceDescriptorTree

pub fn get_config_descriptor(&self, config: u8) -> Result<ConfigDescriptorTree>

Get active config descriptor of this device.

pub fn get_config_descriptor_by_index(
&self,
config_index: u8
) -> Result<ConfigDescriptorTree>

Get a configuration descriptor by its index within the list of descriptors returned by the device.

pub fn get_active_configuration(&self) -> Result<u8>

Get bConfigurationValue of the currently active configuration.

pub fn get_num_configurations(&self) -> u8

Get the total number of configurations for this device.

pub fn clear_halt(&self, ep_addr: u8) -> Result<()>

Clear the halt/stall condition for an endpoint.